+++ R/shapviz_methods.R
@@ -184,3 +184,105 @@ get_shap_interactions.default = function(object, ...) {
   )
 }
 
+#' Dimnames of "shapviz" Object
+#'
+#' @param x An object of class "shapviz".
+#' @return Dimnames of the SHAP matrix.
+#' @export
+#' @examples
+#' S <- matrix(c(1, -1, -1, 1), ncol = 2, dimnames = list(NULL, c("x", "y")))
+#' X <- data.frame(x = c("a", "b"), y = c(100, 10))
+#' x <- shapviz(S, X, baseline = 4)
+#' dimnames(x)
+#'
+#' # Implies colnames()
+#' colnames(x)
+#' @seealso \code{\link{shapviz}}.
+#' @export
+dimnames.shapviz <- function(x) {
+  dimnames(get_shap_values(x))
+}
+
+#' Concatenates "shapviz" Object
+#'
+#' Use standard plus operator to concatenate two "shapviz" objects.
+#'
+#' @param e1 The first object of class "shapviz".
+#' @param e2 The second object of class "shapviz".
+#' @return A new object of class "shapviz".
+#' @export
+#' @examples
+#' S1 <- matrix(c(1, -1, -1, 1), ncol = 2, dimnames = list(NULL, c("x", "y")))
+#' S2 <- matrix(c(-1, 1, 1, -1), ncol = 2, dimnames = list(NULL, c("x", "y")))
+#' X1 <- data.frame(x = c("a", "b"), y = c(100, 10))
+#' X2 <- data.frame(x = c("b", "a"), y = c(100, 10))
+#' x1 <- shapviz(S1, X1, baseline = 4)
+#' x2 <- shapviz(S2, X2, baseline = 4)
+#' x1 + x2
+#' @seealso \code{\link{shapviz}}.
+#' @export
+`+.shapviz` <- function(e1, e2){
+  # input checks
+  stopifnot(
+    is.shapviz(e1),
+    is.shapviz(e2),
+    ncol(e1) == ncol(e2),
+    colnames(e1) == colnames(e2)
+  )
+
+  baseline <- get_baseline(e1)
+  if (baseline != get_baseline(e2)) {
+    warning("Baselines not identical! Will use the one from the first shapviz object.")
+  }
+
+  shapviz(
+    object = rbind(get_shap_values(e1), get_shap_values(e2)),
+    X = rbind(get_feature_values(e1), get_feature_values(e2)),
+    b = baseline,
+    S_inter = rbind_S_inter(get_shap_interactions(e1), get_shap_interactions(e2))
+  )
+}
+
+#' Concatenates "shapviz" Objects
+#'
+#' It is based on the `+.shapviz` operator
+#'
+#' @param ... "shapviz" objects to be concatenated.
+#' @return A new object of class "shapviz".
+#' @export
+#' @examples
+#' S1 <- matrix(c(1, -1, -1, 1), ncol = 2, dimnames = list(NULL, c("x", "y")))
+#' S2 <- matrix(c(-1, 1, 1, -1), ncol = 2, dimnames = list(NULL, c("x", "y")))
+#' X1 <- data.frame(x = c("a", "b"), y = c(100, 10))
+#' X2 <- data.frame(x = c("b", "a"), y = c(100, 10))
+#' x1 <- shapviz(S1, X1, baseline = 4)
+#' x2 <- shapviz(S2, X2, baseline = 4)
+#' rbind(x1, x2)
+#' @seealso \code{\link{shapviz}}.
+#' @export
+rbind.shapviz <- function(...) {
+  Reduce(`+`, list(...))
+}
+
+# Helper functions
+
+# Binds two compatible SHAP interaction arrays along the first dimension.
+rbind_S_inter <- function(x, y) {
+  if (is.null(x) || is.null(y)) {
+    stopifnot(is.null(x) && is.null(y))
+    return(NULL)
+  }
+
+  # Could do many input checks, but consistency is given when called from `+.shapviz`()
+  nx <- nrow(x)
+  out <- array(
+    dim = c(nx + nrow(y), dim(x)[-1L]),
+    dimnames = c(list(NULL), dimnames(x)[-1L])
+  )
+
+  ix <- seq_len(nx)
+  out[ix, , ] <- x
+  out[-ix, , ] <- y
+  out
+}
+