+++ NEWS.md
@@ -1,3 +1,11 @@
+# shapviz 0.7.0
+
+## New features
+
+- Support for "predict_parts" objects from "DALEX" package.
+- "shapviz" objects `x1, x2` can now be concatenated using `x1 + x2` or `rbind(x1, x2)`.
+- "shapviz" objects `x` have received a `dimnames()` function, so you can now, e.g., use `colnames(x)` to see the feature names.
+
 # shapviz 0.6.1
 
 ## New features